Add integration test that fails if Gradle does not understand the JVM's GC settings
