+++ subprojects/launcher/src/main/java/org/gradle/launcher/daemon/server/health/gc/GarbageCollectionMonitor.java
@@ -16,17 +16,12 @@
 
 package org.gradle.launcher.daemon.server.health.gc;
 
-import org.gradle.api.Transformer;
-import org.gradle.api.logging.Logger;
-import org.gradle.api.logging.Logging;
 import org.gradle.api.specs.Spec;
-import org.gradle.internal.jvm.Jvm;
 import org.gradle.internal.time.Time;
 import org.gradle.util.CollectionUtils;
 
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
-import java.util.List;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
@@ -34,16 +29,11 @@ public class GarbageCollectionMonitor {
     private static final int POLL_INTERVAL_SECONDS = 1;
     private static final int POLL_DELAY_SECONDS = 1;
     private static final int EVENT_WINDOW = 20;
-    private static final Logger LOGGER = Logging.getLogger(GarbageCollectionMonitor.class);
 
     private final SlidingWindow<GarbageCollectionEvent> heapEvents;
     private final SlidingWindow<GarbageCollectionEvent> nonHeapEvents;
     private final GarbageCollectorMonitoringStrategy gcStrategy;
     private final ScheduledExecutorService pollingExecutor;
-
-    public GarbageCollectionMonitor(ScheduledExecutorService pollingExecutor) {
-        this(determineGcStrategy(), pollingExecutor);
-    }
 
     public GarbageCollectionMonitor(GarbageCollectorMonitoringStrategy gcStrategy, ScheduledExecutorService pollingExecutor) {
         this.pollingExecutor = pollingExecutor;
@@ -52,29 +42,6 @@ public class GarbageCollectionMonitor {
         this.nonHeapEvents = new DefaultSlidingWindow<GarbageCollectionEvent>(EVENT_WINDOW);
         if (gcStrategy != GarbageCollectorMonitoringStrategy.UNKNOWN) {
             pollForValues();
-        }
-    }
-
-    private static GarbageCollectorMonitoringStrategy determineGcStrategy() {
-
-        final List<String> garbageCollectors = CollectionUtils.collect(ManagementFactory.getGarbageCollectorMXBeans(), new Transformer<String, GarbageCollectorMXBean>() {
-            @Override
-            public String transform(GarbageCollectorMXBean garbageCollectorMXBean) {
-                return garbageCollectorMXBean.getName();
-            }
-        });
-        GarbageCollectorMonitoringStrategy gcStrategy = CollectionUtils.findFirst(GarbageCollectorMonitoringStrategy.values(), new Spec<GarbageCollectorMonitoringStrategy>() {
-            @Override
-            public boolean isSatisfiedBy(GarbageCollectorMonitoringStrategy strategy) {
-                return garbageCollectors.contains(strategy.getGarbageCollectorName());
-            }
-        });
-
-        if (gcStrategy == null) {
-            LOGGER.info("Unable to determine a garbage collection monitoring strategy for " + Jvm.current().toString());
-            return GarbageCollectorMonitoringStrategy.UNKNOWN;
-        } else {
-            return gcStrategy;
         }
     }
 
@@ -94,9 +61,5 @@ public class GarbageCollectionMonitor {
 
     public GarbageCollectionStats getNonHeapStats() {
         return GarbageCollectionStats.forNonHeap(nonHeapEvents.snapshot());
-    }
-
-    public GarbageCollectorMonitoringStrategy getGcStrategy() {
-        return gcStrategy;
     }
 }
