+++ subprojects/launcher/src/main/java/org/gradle/launcher/daemon/server/health/gc/GarbageCollectorMonitoringStrategy.java
@@ -16,13 +16,28 @@
 
 package org.gradle.launcher.daemon.server.health.gc;
 
+import org.gradle.api.Transformer;
+import org.gradle.api.logging.Logger;
+import org.gradle.api.logging.Logging;
+import org.gradle.api.specs.Spec;
+import org.gradle.internal.jvm.Jvm;
+import org.gradle.util.CollectionUtils;
+
+import java.lang.management.GarbageCollectorMXBean;
+import java.lang.management.ManagementFactory;
+import java.lang.management.MemoryPoolMXBean;
+import java.util.List;
+
 public enum GarbageCollectorMonitoringStrategy {
+
     ORACLE_PARALLEL_CMS("PS Old Gen", "Metaspace", "PS MarkSweep", 1.2, 80, 80, 5.0),
     ORACLE_6_CMS("CMS Old Gen", "Metaspace", "ConcurrentMarkSweep", 1.2, 80, 80, 5.0),
     ORACLE_SERIAL("Tenured Gen", "Metaspace", "MarkSweepCompact", 1.2, 80, 80, 5.0),
     ORACLE_G1("G1 Old Gen", "Metaspace", "G1 Old Generation", 0.4, 75, 80, 2.0),
     IBM_ALL("Java heap", "Not Used", "MarkSweepCompact", 0.8, 70, -1, 6.0),
     UNKNOWN(null, null, null, -1, -1, -1, -1);
+
+    private static final Logger LOGGER = Logging.getLogger(GarbageCollectionMonitor.class);
 
     private final String heapPoolName;
     private final String nonHeapPoolName;
@@ -68,5 +83,37 @@ public enum GarbageCollectorMonitoringStrategy {
 
     public double getThrashingThreshold() {
         return thrashingThreshold;
+    }
+
+    public static GarbageCollectorMonitoringStrategy determineGcStrategy() {
+        final List<String> garbageCollectors = CollectionUtils.collect(ManagementFactory.getGarbageCollectorMXBeans(), new Transformer<String, GarbageCollectorMXBean>() {
+            @Override
+            public String transform(GarbageCollectorMXBean garbageCollectorMXBean) {
+                return garbageCollectorMXBean.getName();
+            }
+        });
+        GarbageCollectorMonitoringStrategy gcStrategy = CollectionUtils.findFirst(GarbageCollectorMonitoringStrategy.values(), new Spec<GarbageCollectorMonitoringStrategy>() {
+            @Override
+            public boolean isSatisfiedBy(GarbageCollectorMonitoringStrategy strategy) {
+                return garbageCollectors.contains(strategy.getGarbageCollectorName());
+            }
+        });
+
+        if (gcStrategy == null) {
+            LOGGER.info("Unable to determine a garbage collection monitoring strategy for " + Jvm.current().toString());
+            return GarbageCollectorMonitoringStrategy.UNKNOWN;
+        } else {
+            List<String> memoryPools = CollectionUtils.collect(ManagementFactory.getMemoryPoolMXBeans(), new Transformer<String, MemoryPoolMXBean>() {
+                @Override
+                public String transform(MemoryPoolMXBean memoryPoolMXBean) {
+                    return memoryPoolMXBean.getName();
+                }
+            });
+            if (!memoryPools.contains(gcStrategy.heapPoolName) || !memoryPools.contains(gcStrategy.nonHeapPoolName)) {
+                LOGGER.info("Unable to determine which memory pools to monitor for " + Jvm.current().toString());
+                return GarbageCollectorMonitoringStrategy.UNKNOWN;
+            }
+            return gcStrategy;
+        }
     }
 }
